First test of segments, fits needle OK but connector needs work
Too hard to print a butterfly connector so small, same with
the current V shape. And for the V the angle was additionally
too shallow to actually provide a good lock.
May have to go with M/F connectors printed into the segments
themselves.
Or maybe one can print an U shape in XY plane,
which then fits down+upwards into the segments in square holes.

Another alternative is to have top/bottom layers that
overlap and lock it into place.

Redesigned to be 3x6=18 needles, for first prototypes